Extract useDebouncedInput to inserter hooks
